Global: expands pid to i64
See `process.rs` for detail

+++ src/columns/cgroup.rs
@@ -6,8 +6,8 @@ use std::collections::HashMap;
 pub struct Cgroup {
     header: String,
     unit: String,
-    fmt_contents: HashMap<i32, String>,
-    raw_contents: HashMap<i32, String>,
+    fmt_contents: HashMap<i64, String>,
+    raw_contents: HashMap<i64, String>,
     width: usize,
 }
