+++ src/columns/gid_real.rs
@@ -6,8 +6,8 @@ use std::collections::HashMap;
 pub struct GidReal {
     header: String,
     unit: String,
-    fmt_contents: HashMap<i32, String>,
-    raw_contents: HashMap<i32, u32>,
+    fmt_contents: HashMap<i64, String>,
+    raw_contents: HashMap<i64, u32>,
     width: usize,
 }
