+++ src/columns/pgid.rs
@@ -6,8 +6,8 @@ use std::collections::HashMap;
 pub struct Pgid {
     header: String,
     unit: String,
-    fmt_contents: HashMap<i32, String>,
-    raw_contents: HashMap<i32, i32>,
+    fmt_contents: HashMap<i64, String>,
+    raw_contents: HashMap<i64, i32>,
     width: usize,
 }
