+++ src/columns/uid_login.rs
@@ -8,8 +8,8 @@ const UID_NOT_SET: u32 = 0xffffffff;
 pub struct UidLogin {
     header: String,
     unit: String,
-    fmt_contents: HashMap<i32, String>,
-    raw_contents: HashMap<i32, u32>,
+    fmt_contents: HashMap<i64, String>,
+    raw_contents: HashMap<i64, u32>,
     width: usize,
 }
