+++ src/columns/user.rs
@@ -10,8 +10,8 @@ use uzers::Users;
 pub struct User {
     header: String,
     unit: String,
-    fmt_contents: HashMap<i32, String>,
-    raw_contents: HashMap<i32, String>,
+    fmt_contents: HashMap<i64, String>,
+    raw_contents: HashMap<i64, String>,
     width: usize,
     #[allow(dead_code)]
     abbr_sid: bool,
